+++ R/socket.R
@@ -189,3 +189,45 @@ close.nanoSocket <- function(con, ...) invisible(.Call(rnng_close, con))
 #' @export
 #'
 reap <- function(con) .Call(rnng_reap, con)
+
+#' Monitor a Socket for Pipe Changes
+#'
+#' This function monitors pipe additions and removals from a socket.
+#'
+#' @param sock a Socket.
+#' @param cv a conditionVariable.
+#'
+#' @return For \code{monitor}: a Monitor (object of class 'nanoMonitor'). \cr
+#'   For \code{read_monitor}: an integer vector of pipe IDs (positive if added,
+#'   negative if removed), or else NULL if there were no changes since the
+#'   previous read.
+#'
+#' @examples
+#' cv <- cv()
+#' s <- socket("poly")
+#' s1 <- socket("poly")
+#'
+#' m <- monitor(s, cv)
+#' m
+#'
+#' listen(s)
+#' dial(s1)
+#'
+#' cv_value(cv)
+#' read_monitor(m)
+#'
+#' close(s)
+#' close(s1)
+#'
+#' read_monitor(m)
+#'
+#' @export
+#'
+monitor <- function(sock, cv) .Call(rnng_monitor_create, sock, cv)
+
+#' @param x an external pointer to a monitor.
+#'
+#' @rdname monitor
+#' @export
+#'
+read_monitor <- function(x) .Call(rnng_monitor_read, x)